+++ HaselDebug/Config/PluginConfig.cs
@@ -70,6 +70,7 @@ public partial class PluginConfig
 {
     public int Version { get; set; } = CURRENT_CONFIG_VERSION;
     public bool AutoOpenPluginWindow = false;
+    public bool EnableLuaLogger = false;
     public string LastSelectedTab = "";
     public string[] PinnedInstances = [];
     public bool Excel2Tab_ShowRawSheets = false;