+++ HaselDebug/Windows/ConfigWindow.cs
@@ -1,4 +1,5 @@
 using HaselDebug.Config;
+using HaselDebug.Services;
 
 namespace HaselDebug.Windows;
 
@@ -32,7 +33,20 @@ public partial class ConfigWindow : SimpleWindow
         // AutoOpenPluginWindow
         configChanged |= ImGui.Checkbox($"{_textService.Translate("Config.AutoOpenPluginWindow.Label")}##AutoOpenPluginWindow", ref _pluginConfig.AutoOpenPluginWindow);
 
+        // EnableLuaLogger
+        configChanged |= ImGui.Checkbox($"{_textService.Translate("Config.EnableLuaLogger.Label")}##EnableLuaLogger", ref _pluginConfig.EnableLuaLogger);
+
         if (configChanged)
+        {
             _pluginConfig.Save();
+
+            if (ServiceLocator.TryGetService<LuaLogger>(out var luaLogger))
+            {
+                if (_pluginConfig.EnableLuaLogger)
+                    luaLogger.Enable();
+                else
+                    luaLogger.Disable();
+            }
+        }
     }
 }